### 2025-01-17 #7686 @editorialbot
[PRE REVIEW]: ACEngine: Augmented Reality UNIX C++ Engine for Enhanced Visual Guidance in Digital Fabrication
**Submitting author:** @9and3 (<a href="http://orcid.org/0000-0001-5020-7331">Andrea Settimi</a>)
**Repository:** https://github.com/ibois-epfl/augmented-carpentry
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
