+++ docs/developer-guide/acim.md
@@ -1,0 +1,138 @@
+---
+tags:
+  - acim
+  - I/O
+  - backend
+  - Rhino
+---
+
+# Execution model
+
+The `.acim` file represents the execution model in the **AC** system. An execution model is needed to provide the necessary information for the fabrication process. It can be produced with any 3D modeling software.
+
+`.acim` is the AC's XML-based format, similar to standard exchange formats such as *BTLx*, and contains both geometric information (e.g., joints and holes) and fabrication metadata (e.g., ID, fabrication status).
+
+During our prototyping phase, we use [*Rhinoceros 3D* with a custom script](https://github.com/ibois-epfl/augmented-carpentry/blob/main/py/components/acimexporter/code.py) that automatically detects joints and holes in the 3D model. With a single click, each element (e.g., beam, panel) can be exported to an `.acim` file without manually creating additional 2D documentation.
+
+![acim export image ><](../../assets/images/developer_guide/only_acim_exporter.png){width=500}
+
+
+## `.acim` file structure
+
+The `.acim` file is currently storing all the information required for the fabrication process. The file parse and store the geometries in 3 components:
+
+* `<timber>`: as the main container object
+* `<hole>`: represents all perforations in the model
+* `<cut>`: represents all the cuts in the model
+
+### `<timber>`
+
+The `<timber>` tag represents the main component of the model. It contains the ID, the fabrication status, the current cut, and the bounding box of the component.
+
+The `<bbox>` tag contains the coordinates of the corners of the object oriented bounding box in which the timber element is exported. The corners are represented by their coordinates in the 3D space. This is a fundamental element that we already include in the `.acim` file that will be later used by AC for registering the model to the map (see [`void LayerModel::AlignModels()`](https://github.com/ibois-epfl/augmented-carpentry/blob/ecc39246c4996833e271c96de244561f85d6dc81/src/AIAC/LayerModel.cpp#L121-L179))
+
+[:octicons-file-code-16: `example.acim`](../../assets/example_files/C003.acim)
+```xml
+<timber id="C003">
+  <executed>NotDone</executed>
+  <current>Cut#1</current>
+  <bbox>
+    <corner id="0">-0.0010000377932458662 9.996989965467265e-05 9.989726472193752e-05</corner>
+    <corner id="1">2.296127411397951 9.996989965467265e-05 9.989726472193752e-05</corner>
+    <corner id="2">2.296127411397951 0.13990003039836885 9.989726472193752e-05</corner>
+    <corner id="3">-0.0010000377932458662 0.13990003039836885 9.989726472193752e-05</corner>
+    <corner id="4">-0.0010000377932458662 9.996989965467265e-05 0.13990000059604646</corner>
+    <corner id="5">2.296127411397951 9.996989965467265e-05 0.13990000059604646</corner>
+    <corner id="6">2.296127411397951 0.13990003039836885 0.13990000059604646</corner>
+    <corner id="7">-0.0010000377932458662 0.13990003039836885 0.13990000059604646</corner>
+  </bbox>
+```
+
+
+### `<hole>`
+
+Holes are represented by their vector and radius and have the `<state>` tag to indicate their fabrication status. The `<neighbors>` tag indicates the number of neighboring holes for composite holes (e.g. a hole and a washer). The two following highlighted holes are neighbors, whereas the third hole is not.^(`<neighbors>-1</neighbors>`)
+
[…]
