+++ docs/augmented-woodworking/index.md
@@ -1,1 +1,4 @@
 Placeholder for augmented woodworking
+
+
+![image_test_interface](../assets/images/ac_illustrations/test_img.png){width="700" class="rounded-corners"}
