+++ mkdocs.yml
@@ -111,20 +111,20 @@ plugins:
   - macros:
       module_name: docs/scripts/main
   # FIXME: >>> DOXY >>> uncomment before release
-  - mkdoxy:
-      enabled: !ENV [ENABLE_MKDOXY, True]
-      projects:
-        acdoxygen:
-          src-dirs: src
-          full-doc: True
-          doxy-cfg:
-            FILE_PATTERNS: "*.h* *.c *.cpp *.cc"
-            RECURSIVE: True
-      # save-api: docs/.mkdoxy  # to output the .md files
-      full-doc: True
-      debug: False
-      ignore-errors: False
-      emojis-enabled: True
+  # - mkdoxy:
+  #     enabled: !ENV [ENABLE_MKDOXY, True]
+  #     projects:
+  #       acdoxygen:
+  #         src-dirs: src
+  #         full-doc: True
+  #         doxy-cfg:
+  #           FILE_PATTERNS: "*.h* *.c *.cpp *.cc"
+  #           RECURSIVE: True
+  #     # save-api: docs/.mkdoxy  # to output the .md files
+  #     full-doc: True
+  #     debug: False
+  #     ignore-errors: False
+  #     emojis-enabled: True
   # FIXME: <<< DOXY <<< uncomment before release
 
 
@@ -230,26 +230,26 @@ nav:
       - changelog/index.md
 
   - Augmented Woodworking:
-    - augmented-woodworking/404.md  # TODO: replace me when done
-    # - augmented-woodworking/index.md
+    # - augmented-woodworking/404.md  # TODO: replace me when done
+    - augmented-woodworking/index.md
     - test1: augmented-woodworking/placeholder1.md
     - test2: augmented-woodworking/placeholder2.md
     - Camera calibration: augmented-woodworking/camera-calibration.md
 
-    - C++ docs: 
-      - cpp-docs/index.md
     # FIXME: >>> DOXY >>> uncomment before release
-    - Files: "acdoxygen/files.md"
-    - Functions: "acdoxygen/functions.md"
-    - Classes:
-        - Class List: "acdoxygen/annotated.md"
-        - Class Hierarchy: "acdoxygen/hierarchy.md"
-        - Class Members: "acdoxygen/class_members.md"
-        - Class Member Functions: "acdoxygen/class_member_functions.md"
-        - Class Member Variables: "acdoxygen/class_member_variables.md"
+    # - C++ docs: 
+    #   - cpp-docs/index.md
+    # - Files: "acdoxygen/files.md"
+    # - Functions: "acdoxygen/functions.md"
+    # - Classes:
+    #     - Class List: "acdoxygen/annotated.md"
+    #     - Class Hierarchy: "acdoxygen/hierarchy.md"
+    #     - Class Members: "acdoxygen/class_members.md"
+    #     - Class Member Functions: "acdoxygen/class_member_functions.md"
+    #     - Class Member Variables: "acdoxygen/class_member_variables.md"
+    #     - Variables: "acdoxygen/variables.md"
+    # - Macros: "acdoxygen/macros.md"
     # FIXME: <<< DOXY <<< uncomment before release
-        - Variables: "acdoxygen/variables.md"
-    - Macros: "acdoxygen/macros.md"
 
   - Developer guide:
     - AR engine:
@@ -258,6 +258,7 @@ nav:
       - Feedback system: developer-guide/AR-engine/feedback-system.md
       - AR rendering: developer-guide/AR-engine/AR-rendering.md
     - Toolhead: developer-guide/toolhead.md
+    - Execution model: developer-guide/acim.md
 
     - Guidelines:
       - Code: developer-guide/CONTRIBUTING_copy.md  # TODO: add a copy in the githhub action
