### 2025-01-18 89b82b9
FIX: cpp session reintroduced in mkdocs
