### 2025-01-18 e2b249f
WIP-FIX: testing a possible solution for community page
