+++ mkdocs.yml
@@ -264,7 +264,7 @@ nav:
       - Tags: developer-guide/tags.md
   
   - Community:
-    - contributing/index.md
+    - contributing/community.md
     - Reporting a bug: contributing/reporting-a-bug.md
     - Reporting a docs issue: contributing/reporting-a-docs-issue.md
     - Adding tools or mounts: contributing/add-new-tool.md
