### 2025-01-18 0e7e952
FIX: missing image in ar engine docs
