### 2025-01-19 e110dbe
FIX: better image for acim docs
