+++ mkdocs.yml
@@ -111,20 +111,20 @@ plugins:
   - macros:
       module_name: docs/scripts/main
   # FIXME: >>> DOXY >>> uncomment before release
-  - mkdoxy:
-      enabled: !ENV [ENABLE_MKDOXY, True]
-      projects:
-        acdoxygen:
-          src-dirs: src
-          full-doc: True
-          doxy-cfg:
-            FILE_PATTERNS: "*.h* *.c *.cpp *.cc"
-            RECURSIVE: True
-      # save-api: docs/.mkdoxy  # to output the .md files
-      full-doc: True
-      debug: False
-      ignore-errors: False
-      emojis-enabled: True
+  # - mkdoxy:
+  #     enabled: !ENV [ENABLE_MKDOXY, True]
+  #     projects:
+  #       acdoxygen:
+  #         src-dirs: src
+  #         full-doc: True
+  #         doxy-cfg:
+  #           FILE_PATTERNS: "*.h* *.c *.cpp *.cc"
+  #           RECURSIVE: True
+  #     # save-api: docs/.mkdoxy  # to output the .md files
+  #     full-doc: True
+  #     debug: False
+  #     ignore-errors: False
+  #     emojis-enabled: True
   # FIXME: <<< DOXY <<< uncomment before release
 
 
@@ -236,18 +236,18 @@ nav:
     - Camera calibration: augmented-woodworking/camera-calibration.md
 
     # FIXME: >>> DOXY >>> uncomment before release
-  - C++ docs: 
-    - cpp-docs/index.md
-    - Files: "acdoxygen/files.md"
-    - Functions: "acdoxygen/functions.md"
-    - Classes:
-        - Class List: "acdoxygen/annotated.md"
-        - Class Hierarchy: "acdoxygen/hierarchy.md"
-        - Class Members: "acdoxygen/class_members.md"
-        - Class Member Functions: "acdoxygen/class_member_functions.md"
-        - Class Member Variables: "acdoxygen/class_member_variables.md"
-        - Variables: "acdoxygen/variables.md"
-    - Macros: "acdoxygen/macros.md"
+  # - C++ docs: 
+  #   - cpp-docs/index.md
+  #   - Files: "acdoxygen/files.md"
+  #   - Functions: "acdoxygen/functions.md"
+  #   - Classes:
+  #       - Class List: "acdoxygen/annotated.md"
+  #       - Class Hierarchy: "acdoxygen/hierarchy.md"
+  #       - Class Members: "acdoxygen/class_members.md"
+  #       - Class Member Functions: "acdoxygen/class_member_functions.md"
+  #       - Class Member Variables: "acdoxygen/class_member_variables.md"
+  #       - Variables: "acdoxygen/variables.md"
+  #   - Macros: "acdoxygen/macros.md"
     # FIXME: <<< DOXY <<< uncomment before release
 
   - Developer guide:
