+++ docs/developer-guide/toolheads.md
@@ -1,8 +1,0 @@
----
-tags:
-  - toolheads
----
-
-# Toolheads
-
-Here we give more information about how toolheads are itnegrated in the AC system.
