### 2025-01-23 164228f
ADD: extra UI image for acim docs
