### 2025-01-27 3e376e0
ADD: temporary example file for cutting image
