+++ docs/augmented-woodworking/index.md
@@ -1,4 +1,6 @@
 Placeholder for augmented woodworking
 
 
+![image_test_interface](../assets/images/test/placeholder_cut.png){width="700" class="rounded-corners"}
+
 ![image_test_interface](../assets/images/ac_illustrations/test_img.png){width="700" class="rounded-corners"}
