+++ assets/.doxy/acdoxygen/acdoxygen/HoleToolheadAxisExporter_8cpp_source.md
@@ -1,0 +1,129 @@
+
+
+# File HoleToolheadAxisExporter.cpp
+
+[**File List**](files.md) **>** [**src**](dir_68267d1309a1af8e8297ef4c3efbcdba.md) **>** [**utils**](dir_313caf1132e152dd9b58bea13a4052ca.md) **>** [**HoleToolheadAxisExporter.cpp**](HoleToolheadAxisExporter_8cpp.md)
+
+[Go to the documentation of this file](HoleToolheadAxisExporter_8cpp.md)
+
+
+```C++
+// #####################################################################
+// >>>>>>>>>>>>>>>>>>>>> BEGINNING OF LEGAL NOTICE >>>>>>>>>>>>>>>>>>>>>
+//######################################################################
+//
+// This source file, along with its associated content, was authored by
+// Andrea Settimi, Hong-Bin Yang, Naravich Chutisilp, and numerous other
+// contributors. The code was originally developed at the Laboratory for
+// Timber Construction (IBOIS, director: Prof. Yves Weinand) at the School of 
+// Architecture, Civil and Environmental Engineering (ENAC) at the Swiss
+// Federal Institute of Technology in Lausanne (EPFL) for the Doctoral
+// Research "Augmented Carpentry" (PhD researcher: Andrea Settimi,
+// co-director: Dr. Julien Gamerro, director: Prof. Yves Weinand).
+//
+// Although the entire repository is distributed under the GPL license,
+// these particular source files may also be used under the terms of the
+// MIT license. By accessing or using this file, you agree to the following:
+//
+// 1. You may reproduce, modify, and distribute this file in accordance
+//    with the terms of the MIT license.
+// 2. You must retain this legal notice in all copies or substantial
+//    portions of this file.
+// 3. This file is provided "AS IS," without any express or implied
+//    warranties, including but not limited to the implied warranties of
+//    merchantability and fitness for a particular purpose.
+//
+// If you cannot or will not comply with the above conditions, you are
+// not permitted to use this file. By proceeding, you acknowledge and
+// accept all terms and conditions herein.
+//
+//######################################################################
+// <<<<<<<<<<<<<<<<<<<<<<< END OF LEGAL NOTICE <<<<<<<<<<<<<<<<<<<<<<<<
+// #####################################################################
+//
+#include "HoleToolheadAxisExporter.h"
+#include "AIAC/ACInfoModel.h"
+#include "AIAC/Application.h"
+#include "utils/VideoRecorder.h"
+#include "AIAC/LayerUtils.h"
+
+namespace AIAC::Utils {
+
+    HoleToolheadAxisExporter::HoleToolheadAxisExporter(const std::string& basePath)
+    : m_BasePath(basePath)
+    {
+        LayerUtils::CreateFolder(this->m_BasePath);
+    };
+
+    void HoleToolheadAxisExporter::ExportCoordinates(){
+        this->ExportToolheadAxis();
+        this->ExportHoleAxis();
+        this->WriteBufferToFile();
+    }
+
+    void HoleToolheadAxisExporter::ExportToolheadAxis(){
+        ACToolHeadType activeToolheadType = AC_FF_TOOL->GetType();
+        if (activeToolheadType == ACToolHeadType::DRILLBIT){
+            auto drillBitData = AC_FF_TOOL->GetData<DrillBitData>();
+            this->WriteCoordToBuffer("Toolhead", AC_FF_TOOL->GetName(), "ToolbaseGO", drillBitData.ToolbaseGO);
[…]
