+++ assets/.doxy/acdoxygen/acdoxygen/LayerToolhead_8cpp_source.md
@@ -1,0 +1,243 @@
+
+
+# File LayerToolhead.cpp
+
+[**File List**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**LayerToolhead.cpp**](LayerToolhead_8cpp.md)
+
+[Go to the documentation of this file](LayerToolhead_8cpp.md)
+
+
+```C++
+// #####################################################################
+// >>>>>>>>>>>>>>>>>>>>> BEGINNING OF LEGAL NOTICE >>>>>>>>>>>>>>>>>>>>>
+//######################################################################
+//
+// This source file, along with its associated content, was authored by
+// Andrea Settimi, Hong-Bin Yang, Naravich Chutisilp, and numerous other
+// contributors. The code was originally developed at the Laboratory for
+// Timber Construction (IBOIS, director: Prof. Yves Weinand) at the School of 
+// Architecture, Civil and Environmental Engineering (ENAC) at the Swiss
+// Federal Institute of Technology in Lausanne (EPFL) for the Doctoral
+// Research "Augmented Carpentry" (PhD researcher: Andrea Settimi,
+// co-director: Dr. Julien Gamerro, director: Prof. Yves Weinand).
+//
+// Although the entire repository is distributed under the GPL license,
+// these particular source files may also be used under the terms of the
+// MIT license. By accessing or using this file, you agree to the following:
+//
+// 1. You may reproduce, modify, and distribute this file in accordance
+//    with the terms of the MIT license.
+// 2. You must retain this legal notice in all copies or substantial
+//    portions of this file.
+// 3. This file is provided "AS IS," without any express or implied
+//    warranties, including but not limited to the implied warranties of
+//    merchantability and fitness for a particular purpose.
+//
+// If you cannot or will not comply with the above conditions, you are
+// not permitted to use this file. By proceeding, you acknowledge and
+// accept all terms and conditions herein.
+//
+//######################################################################
+// <<<<<<<<<<<<<<<<<<<<<<< END OF LEGAL NOTICE <<<<<<<<<<<<<<<<<<<<<<<<
+// #####################################################################
+//
+#include "aiacpch.h"
+
+#include "AIAC/Config.h"
+
+#include "AIAC/Log.h"
+#include "AIAC/Application.h"
+#include "AIAC/LayerToolhead.h"
+#include "GeometryUtils.h"
+#include "AIAC/UI/ClrPalette.h"
+
+namespace AIAC
+{
+    void LayerToolhead::OnAttach()
+    {
+        bool initDraw = true;
+
+#ifdef HEADLESS_TEST
+        initDraw = false;
+#endif
+
+        // init ttool
+        TTool = std::make_shared<ttool::TTool>(
+            AIAC::Config::Get<std::string>(AIAC::Config::SEC_TTOOL, AIAC::Config::TTOOL_ROOT_PATH, "deps/TTool"),
+            AIAC::Config::Get<std::string>(AIAC::Config::SEC_TTOOL, AIAC::Config::CONFIG_FILE, "deps/TTool/assets/config.yml"),
+            AIAC::Config::Get<std::string>(AIAC::Config::SEC_AIAC, AIAC::Config::CAM_PARAMS_FILE, "assets/tslam/calibration_orange_A_1280_720_000B.yml"),
+            initDraw
+            );
+
+        // load the datasets acits
[…]
