+++ assets/.doxy/acdoxygen/acdoxygen/glob_8cpp.md
@@ -1,0 +1,94 @@
+
+
+# File glob.cpp
+
+
+
+[**FileList**](files.md) **>** [**src**](dir_68267d1309a1af8e8297ef4c3efbcdba.md) **>** [**utils**](dir_313caf1132e152dd9b58bea13a4052ca.md) **>** [**glob.cpp**](glob_8cpp.md)
+
+[Go to the source code of this file](glob_8cpp_source.md)
+
+
+
+* `#include "glob.h"`
+* `#include <dirent.h>`
+* `#include <fnmatch.h>`
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Namespaces
+
+| Type | Name |
+| ---: | :--- |
+| namespace | [**glob**](namespaceglob.md) <br> |
+
+
+## Classes
+
+| Type | Name |
+| ---: | :--- |
+| class | [**glob\_impl**](classglob_1_1glob__impl.md) <br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+------------------------------
+The documentation for this class was generated from the following file `src/utils/glob.cpp`
+
