+++ assets/.doxy/acdoxygen/acdoxygen/LayerModel_8cpp.md
@@ -1,0 +1,94 @@
+
+
+# File LayerModel.cpp
+
+
+
+[**FileList**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**LayerModel.cpp**](LayerModel_8cpp.md)
+
+[Go to the source code of this file](LayerModel_8cpp_source.md)
+
+
+
+* `#include "aiacpch.h"`
+* `#include "Config.h"`
+* `#include "AIAC/LayerModel.h"`
+* `#include "AIAC/GOSys/GO.h"`
+* `#include "AIAC/Application.h"`
+* `#include "AIAC/DLoader.h"`
+* `#include "GeometryUtils.h"`
+* `#include "glm/gtx/string_cast.hpp"`
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Namespaces
+
+| Type | Name |
+| ---: | :--- |
+| namespace | [**AIAC**](namespaceAIAC.md) <br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+------------------------------
+The documentation for this class was generated from the following file `src/AIAC/LayerModel.cpp`
+

[…]
