### 2025-02-27 86a1681
FIX: tiny typo in paper
