+++ docs/augmented-woodworking/map.md
@@ -1,0 +1,77 @@
+---
+tags:
+  - map
+  - UI
+  - fab
+---
+
+<!-- FIXME: Get rid of the title in the page? -->
+
+Before you map the piece, stick the tags
+
+<!-- TODO: Add tags to the timber piece -->
+
+Go around the piece to map it
+
+<!-- TODO: Add the video side by side of the mapping of one piece (maybe with drone??) -->
+
+
+<!-- Start the mapping -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/map_selectstart_01.png)
+</figure>
+<div class="description" markdown>
+Click the button **Save & Exit** (**a**) to start the mapping process.
+</div>
+</div>
+
+<!-- Mapping interface -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/map_inside_map_02.png)
+</figure>
+<div class="description" markdown>
+The mapping is now started.
+
+**a** - landmark feature points used to facilitate camera localization
+
+**b** - tags already acquired
+
+**c** - newly detected tags pending acquisition
+
+**d** - map-building overview
+
+**e** - optimization parameters
+
+**f** - reconstruction parameters for generating a mesh box from the registered tags
+
+**g** - commands to save or discard the current mapping recording
+</div>
+</div>
+
+<!-- Mid mapping -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/map_midmap_03b.png)
+</figure>
+<div class="description" markdown>
+Go around the timber piece, turn it around, and capture as many tags as possible.
+</div>
+</div>
+
+<!-- Mapping closing -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/map_confirm_map_03.png)
+</figure>
+<div class="description" markdown>
+Click on the **Save & Exit** button to save the mapping and start the processing.
+</div>
+</div>
+
+<!-- Load acim model (UI interface to select acim navigation) -->
+
+<!-- Visualize the piece with 3D model overlay -->
+
+<!-- *Extra: Custom reconstruction -->
