+++ docs/augmented-woodworking/camera-calibration.md
@@ -1,8 +1,0 @@
-# Camera calibration
-
-Test
-
-## How to calibrate the camera
-
-
-To be filled
