+++ README.md
@@ -10,7 +10,13 @@
 
 
 This is the repo containing the necessary code to run the **Augmented Carpentry (AC)** software. This is a research thesis project from the [Laboratory of Timber Constructions (IBOIS)](https://ibois.epfl.ch/) at the [Swiss Federal Institute of Technology in Lausanne (EPFL)](https://www.epfl.ch/). It is an Augmented Reality (AR) application that helps operators to integrate ordinary tools (e.g. power drills, circular saws and chainsaws) into digitally guided woodworking to obtain live computed feedback. The aim is to suppress 2D execution drawings and markings and empower the worker with precise and real-time guidance to replicate even a complex digital designs with manual tools.
-Here's an overview of the hardware system for AC. Visit the documentation for more info and instructions, [here](https://ibois-epfl.github.io/augmented-carpentry/):
+Here's a capture from the AC system in action. Visit the documentation for more info and instructions, [here](https://ibois-epfl.github.io/augmented-carpentry/):
+
+
+<p align="center">
+    <img src="https://github.com/user-attachments/assets/007a47d5-cb45-478f-b9c8-42a9deddf905" width="800px">
+</p>
+
 
 #### Computing unit, interface and sensors
 AC works on *x64 machines* running on *Ubuntu 22.04 LTS* and a simple *monocular camera*. For a quick start, you can use a laptop and your webcam. If you want to build on a touch screen, you can use the WaveShare WS170120 and follow the instructions below:
