### 2025-03-13 c62472d
FIX: adjust readme img size
