+++ assets/.doxy/acdoxygen/acdoxygen/GeometryUtils_8cpp_source.md
@@ -1,0 +1,164 @@
+
+
+# File GeometryUtils.cpp
+
+[**File List**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**GeometryUtils.cpp**](GeometryUtils_8cpp.md)
+
+[Go to the documentation of this file](GeometryUtils_8cpp.md)
+
+
+```C++
+// #####################################################################
+// >>>>>>>>>>>>>>>>>>>>> BEGINNING OF LEGAL NOTICE >>>>>>>>>>>>>>>>>>>>>
+//######################################################################
+//
+// This source file, along with its associated content, was authored by
+// Andrea Settimi, Hong-Bin Yang, Naravich Chutisilp, and numerous other
+// contributors. The code was originally developed at the Laboratory for
+// Timber Construction (IBOIS, director: Prof. Yves Weinand) at the School of 
+// Architecture, Civil and Environmental Engineering (ENAC) at the Swiss
+// Federal Institute of Technology in Lausanne (EPFL) for the Doctoral
+// Research "Augmented Carpentry" (PhD researcher: Andrea Settimi,
+// co-director: Dr. Julien Gamerro, director: Prof. Yves Weinand).
+//
+// Although the entire repository is distributed under the GPL license,
+// these particular source files may also be used under the terms of the
+// MIT license. By accessing or using this file, you agree to the following:
+//
+// 1. You may reproduce, modify, and distribute this file in accordance
+//    with the terms of the MIT license.
+// 2. You must retain this legal notice in all copies or substantial
+//    portions of this file.
+// 3. This file is provided "AS IS," without any express or implied
+//    warranties, including but not limited to the implied warranties of
+//    merchantability and fitness for a particular purpose.
+//
+// If you cannot or will not comply with the above conditions, you are
+// not permitted to use this file. By proceeding, you acknowledge and
+// accept all terms and conditions herein.
+//
+//######################################################################
+// <<<<<<<<<<<<<<<<<<<<<<< END OF LEGAL NOTICE <<<<<<<<<<<<<<<<<<<<<<<<
+// #####################################################################
+//
+#include "GeometryUtils.h"
+
+namespace AIAC{
+    glm::mat3x3 GetRotationMatrix(glm::vec3 axis, float theta){
+        // normalize axis
+        axis = glm::normalize(axis);
+
+        float a = cos(theta / 2.0f);
+        float b = -axis.x * sin(theta / 2.0f);
+        float c = -axis.y * sin(theta / 2.0f);
+        float d = -axis.z * sin(theta / 2.0f);
+        float aa = a * a;
+        float bb = b * b;
+        float cc = c * c;
+        float dd = d * d;
+        float bc = b * c;
+        float ad = a * d;
+        float ac = a * c;
+        float ab = a * b;
+        float bd = b * d;
+        float cd = c * d;
+
+        glm::mat3x3 mat;
+        mat[0][0] = aa + bb - cc - dd;
+        mat[0][1] = 2 * (bc + ad);
+        mat[0][2] = 2 * (bd - ac);
+        mat[1][0] = 2 * (bc - ad);
+        mat[1][1] = aa + cc - bb - dd;
+        mat[1][2] = 2 * (cd + ab);
+        mat[2][0] = 2 * (bd + ac);
+        mat[2][1] = 2 * (cd - ab);
[…]
