+++ assets/.doxy/acdoxygen/acdoxygen/Assert_8h_source.md
@@ -1,0 +1,65 @@
+
+
+# File Assert.h
+
+[**File List**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**Assert.h**](Assert_8h.md)
+
+[Go to the documentation of this file](Assert_8h.md)
+
+
+```C++
+// #####################################################################
+// >>>>>>>>>>>>>>>>>>>>> BEGINNING OF LEGAL NOTICE >>>>>>>>>>>>>>>>>>>>>
+//######################################################################
+//
+// This source file, along with its associated content, was authored by
+// Andrea Settimi, Hong-Bin Yang, Naravich Chutisilp, and numerous other
+// contributors. The code was originally developed at the Laboratory for
+// Timber Construction (IBOIS, director: Prof. Yves Weinand) at the School of 
+// Architecture, Civil and Environmental Engineering (ENAC) at the Swiss
+// Federal Institute of Technology in Lausanne (EPFL) for the Doctoral
+// Research "Augmented Carpentry" (PhD researcher: Andrea Settimi,
+// co-director: Dr. Julien Gamerro, director: Prof. Yves Weinand).
+//
+// Although the entire repository is distributed under the GPL license,
+// these particular source files may also be used under the terms of the
+// MIT license. By accessing or using this file, you agree to the following:
+//
+// 1. You may reproduce, modify, and distribute this file in accordance
+//    with the terms of the MIT license.
+// 2. You must retain this legal notice in all copies or substantial
+//    portions of this file.
+// 3. This file is provided "AS IS," without any express or implied
+//    warranties, including but not limited to the implied warranties of
+//    merchantability and fitness for a particular purpose.
+//
+// If you cannot or will not comply with the above conditions, you are
+// not permitted to use this file. By proceeding, you acknowledge and
+// accept all terms and conditions herein.
+//
+//######################################################################
+// <<<<<<<<<<<<<<<<<<<<<<< END OF LEGAL NOTICE <<<<<<<<<<<<<<<<<<<<<<<<
+// #####################################################################
+//
+#pragma once
+
+#include "AIAC/Log.h"
+#include "AIAC/Base.h"
+
+#ifdef AIAC_ENABLE_ASSERTS
+// Alteratively we could use the same "default" message for both "WITH_MSG" and "NO_MSG" and
+    // provide support for custom formatting by concatenating the formatting string instead of having the format inside the default message
+    #define AIAC_INTERNAL_ASSERT_IMPL(type, check, msg, ...) { if(!(check)) { (msg, __VA_ARGS__); AIAC_BREAK(); } }
+    #define AIAC_INTERNAL_ASSERT_WITH_MSG(type, check, ...) AIAC_INTERNAL_ASSERT_IMPL(type, check, "Assertion failed: {0}", __VA_ARGS__)
+    #define AIAC_INTERNAL_ASSERT_NO_MSG(type, check) AIAC_INTERNAL_ASSERT_IMPL(type, check, "Assertion '{0}' failed at {1}:{2}", AIAC_STRINGIFY_MACRO(check), std::filesystem::path(__FILE__).filename().string(), __LINE__)
+
+    #define AIAC_INTERNAL_ASSERT_GET_MACRO_NAME(arg1, arg2, macro, ...) macro
[…]
