+++ assets/.doxy/acdoxygen/acdoxygen/Shader_8hpp_source.md
@@ -1,0 +1,19 @@
+
+
+# File Shader.hpp
+
+[**File List**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**Render**](dir_4231f99b70cbd7a69a19f070b3954fcb.md) **>** [**Shader.hpp**](Shader_8hpp.md)
+
+[Go to the documentation of this file](Shader_8hpp.md)
+
+
+```C++
+#ifndef SHADER_HPP
+#define SHADER_HPP
+
+GLuint LoadShaders(const char * vertex_file_path,const char * fragment_file_path);
+
+#endif
+```
+
+
