+++ assets/.doxy/acdoxygen/acdoxygen/ACInfoToolhead_8cpp.md
@@ -1,0 +1,90 @@
+
+
+# File ACInfoToolhead.cpp
+
+
+
+[**FileList**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**ACInfoToolhead.cpp**](ACInfoToolhead_8cpp.md)
+
+[Go to the source code of this file](ACInfoToolhead_8cpp_source.md)
+
+
+
+* `#include "ACInfoToolhead.h"`
+* `#include "AIAC/Log.h"`
+* `#include "GeometryUtils.h"`
+* `#include "pugixml.hpp"`
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Namespaces
+
+| Type | Name |
+| ---: | :--- |
+| namespace | [**AIAC**](namespaceAIAC.md) <br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+------------------------------
+The documentation for this class was generated from the following file `src/AIAC/ACInfoToolhead.cpp`
+

[…]
