### 2025-03-13 6a092d6
CAP-ADD: video template for tutorials with mask
