+++ docs/augmented-woodworking/map.md
@@ -50,6 +50,7 @@ The mapping is now started.
 </div>
 </div>
 
+<!-- TODO: transform this in a GIF and add the template to css -->
 <!-- Mid mapping -->
 <div class="sevinch-display-tutorial" markdown>
 <figure markdown>
@@ -69,6 +70,33 @@ Go around the timber piece, turn it around, and capture as many tags as possible
 Click on the **Save & Exit** button to save the mapping and start the processing.
 </div>
 </div>
+
+<!-- In case of reconstruction failure -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1065422379?h=3461c28610&amp;badge=1&amp;autopause=0&amp;player_id=0&amp;app_id=58479&amp;title=1&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+        class="frame-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+This is a **test** description.
+</div>
+</div>
+
+
+
+
+
+
 
 <!-- Load acim model (UI interface to select acim navigation) -->
 
