+++ docs/augmented-woodworking/map.md
@@ -22,7 +22,7 @@ Go around the piece to map it
 ![Image title](../assets/images/woodworking/map_selectstart_01.png)
 </figure>
 <div class="description" markdown>
-Click the button **Save & Exit** (**a**) to start the mapping process.
+Click the button **Start Mapping** (*) to start the mapping process.
 </div>
 </div>
 
@@ -36,9 +36,9 @@ The mapping is now started.
 
 **a** - landmark feature points used to facilitate camera localization
 
-**b** - tags already acquired
+**b** - tags already acquired (yellow)
 
-**c** - newly detected tags pending acquisition
+**c** - newly detected tags pending acquisition (blue)
 
 **d** - map-building overview
 
@@ -50,14 +50,24 @@ The mapping is now started.
 </div>
 </div>
 
-<!-- TODO: transform this in a GIF and add the template to css -->
-<!-- Mid mapping -->
+<!-- Mid mapping NB: autovideo-->
 <div class="sevinch-display-tutorial" markdown>
 <figure markdown>
-![Image title](../assets/images/woodworking/map_midmap_03b.png)
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1065559795?h=afb4e097b9&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+        class="frame-overlay"
+    />
+</div>
 </figure>
 <div class="description" markdown>
-Go around the timber piece, turn it around, and capture as many tags as possible.
+Go around the timber piece, turn it when needed, and capture as many tags as possible.
 </div>
 </div>
 
@@ -71,6 +81,11 @@ Click on the **Save & Exit** button to save the mapping and start the processing
 </div>
 </div>
 
+<!-- TODO: Load acim model (UI interface to select acim navigation) -->
+
+<!-- TODO: Visualize the piece with 3D model overlay -->
+
+<!-- FIXME: add the tab block inside, see how it can work>
 <!-- In case of reconstruction failure -->
 <div class="sevinch-display-tutorial" markdown>
 <figure markdown>
@@ -88,18 +103,6 @@ Click on the **Save & Exit** button to save the mapping and start the processing
 </div>
 </figure>
 <div class="description" markdown>
-This is a **test** description.
+In case the reconstruction fails, follow **this** procedure. You can try to reconstruct the piece with different parameters. We have several presets available to try based on the element's length.
 </div>
 </div>
-
-
-
-
-
-
-
-<!-- Load acim model (UI interface to select acim navigation) -->
-
-<!-- Visualize the piece with 3D model overlay -->
-
-<!-- *Extra: Custom reconstruction -->
