+++ docs/augmented-woodworking/map.md
@@ -5,16 +5,14 @@ tags:
   - fab
 ---
 
-<!-- FIXME: Get rid of the title in the page? -->
+Print the [tags](../hardware/stickers.md) and stick them to the timber piece. Cover the middle of each face of the timber beam with the tags. Once it is done, mapping can start. The mapping is the process of capturing the tags with the camera to create a 3D model of the timber piece. By doing this, the system will be able to track the timber piece in real-time during the fabrication.
 
-Before you map the piece, stick the tags
+<!-- FIXME: do the round border -->
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1065880051?h=ea7b8858ac&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
 
-<!-- TODO: Add tags to the timber piece -->
+<br>
 
-Go around the piece to map it
-
-<!-- TODO: Add the video side by side of the mapping of one piece (maybe with drone??) -->
-
+# Step-by-step
 
 <!-- Start the mapping -->
 <div class="sevinch-display-tutorial" markdown>
@@ -67,7 +65,7 @@ The mapping is now started.
 </div>
 </figure>
 <div class="description" markdown>
-Go around the timber piece, turn it when needed, and capture as many tags as possible.
+Go around the timber piece, turn it when needed, and capture as many tags as possible. Start from the middle of the beam and go on the extremities, this will help redistribute the error of the tags detection uniformly.
 </div>
 </div>
 
@@ -81,17 +79,11 @@ Click on the **Save & Exit** button to save the mapping and start the processing
 </div>
 </div>
 
-<!-- TODO: Load acim model (UI interface to select acim navigation) -->
-
-<!-- TODO: Visualize the piece with 3D model overlay -->
-
-<!-- FIXME: add the tab block inside, see how it can work>
-<!-- In case of reconstruction failure -->
 <div class="sevinch-display-tutorial" markdown>
 <figure markdown>
 <div class="video-container">
     <iframe 
-        src="https://player.vimeo.com/video/1065422379?h=3461c28610&amp;badge=1&amp;autopause=0&amp;player_id=0&amp;app_id=58479&amp;title=1&amp;byline=0&amp;portrait=0" 
+        src="https://player.vimeo.com/video/1065789261?h=8a4c544793&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
         frameborder="0" 
         allow="autoplay; fullscreen; picture-in-picture" 
         allowfullscreen>
@@ -103,6 +95,51 @@ Click on the **Save & Exit** button to save the mapping and start the processing
 </div>
 </figure>
 <div class="description" markdown>
-In case the reconstruction fails, follow **this** procedure. You can try to reconstruct the piece with different parameters. We have several presets available to try based on the element's length.
+Once the map is loaded together with the `.acim` model, you can visualize if the piece is correctly reconstructed (black lines) and if the model is correctly aligned with the piece.
 </div>
 </div>
+
+<br>
+
+!!! example "Troubleshooting a failed reconstruction"
+    <div class="sevinch-display-tutorial" markdown>
+    <figure markdown>
+    <div class="video-container">
+        <iframe 
[…]
