+++ docs/augmented-woodworking/execution-model.md
@@ -5,7 +5,11 @@ tags:
   - fab
 ---
 
-Description
+Once you mapped the piece, you can load the `.acim` model to lock the execution model to the current timber element. This will allow you to visualize directly in 3D all the augmented cuts, drills and other information directly on the timber piece.
+
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1066589698?h=a4d4abba68&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;" title="3dacim_view_static"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
+
+<br>
 
 # Step-by-step
 
@@ -29,3 +33,152 @@ Description
 Once the mapping is processed and saved, you can load the `.acim` model to lock the execution model to the current piece.
 </div>
 </div>
+
+
+<!-- acim loaded, general overview -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1066631456?h=863728af2c&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+        class="frame-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+The execution model is now locked to the timber element and it can now be visualized in 3D.
+</div>
+</div>
+
+
+<!-- acim loaded, moving piece -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1066595090?h=4f62581601&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+        class="frame-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Since the model is referenced to the timber element, moving the piece will not affect the tracking.
+</div>
+</div>
+
+
+<!-- Overview of the pannel UI + cuts/drill widget on piece -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/acim_lock_overview.png)
+</figure>
+<div class="description" markdown>
+Let's start by looking at the different elements of the ACIM interface when the model is loaded.
+
+**a** - a selected hole
+
+**b** - a lap-joint in the unselected state
+
+**c** - a hole in the unselected state
+
+**d** - a selected lap-joint
+
+**e** - a UI panel for execution model controls (`acim`)
+
+**f** - dimension lines (cotas) that can be toggled on or off
+
+**g** - basic I/O controls for ACIM
+
+**h** - visual cues indicating the bounding box
+
+**i** - a widget delineating the outer boundaries of the imported execution model
+
+**o** - controls for adjusting the model's position and orientation
+
[…]
