+++ docs/augmented-woodworking/tool-calib.md
@@ -1,1 +1,157 @@
-Calibration procedures with ttools
+---
+tags:
+  - calibration
+  - acit
+  - UI
+  - fab
+---
+
+Here we describe the calibration process of the tool heads. This is a crucial step to ensure the correct tracking of the inserted tool. This procedure needs to be done every time a tool is changed.
+
+<!-- Overview mosaic video of the claibration process -->
+
+<br>
+
+# Step-by-step
+
+<!-- Select the tool UI explanation -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/calibtool_selectui.png)
+</figure>
+<div class="description" markdown>
+Open the **Toolhead** pane and select the tool you want to calibrate from the list of available tool heads:
+
+**a** - rough projection of the selected tool head
+
+**b** - library of selectable tool heads models
+
+**c** - current selected tool head
+
+<br>
+
+If you want to add new tools follow the instructions in the section [contributing: add a new tool](../contributing/add-new-tool.md).
+</div>
+</div>
+
+<!-- image: Input the pose manually UI explanation -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/calibtool_inputglobal.png)
+</figure>
+<div class="description" markdown>
+Once you have adjusted the camera to have the tool occupying the first half of the screen, you can input the initial pose of the tool head via the sliders (**q r p l o n**) first and then refine it (**f**):
+
+**d** - machine-refined tool head position and orientation
+
+**e** - confirm the pose
+
+**f** - stop the refiner
+
+**g** - reload the latest saved pose
+
+**h** - hide the 3D model silhouette widget
+
+**i** - reset the pose to the default value
+
+**q r p** - sliders to input the 3D model's location
+
+**l o n** - sliders to input the 3D model's rotation
+
+</div>
+</div>
+
+
+<!-- video-auto: Input the pose manually -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1066887022?h=85c72e225f&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+        class="frame-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Use the sliders to estimate an initial position. It does not need to be precise, but it should be as close as possible to the real position and orientation of the tool head. Once you are satisfied, hit the button **Save pose**.
+</div>
+</div>
+
+
+<!-- video-auto: Refine the pose -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1066892093?h=c8006bc436&amp;badge=0&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
[…]
