+++ docs/augmented-woodworking/draw.md
@@ -1,1 +1,11 @@
-Placeholder
+---
+tags:
+  - design
+  - acim
+  - UI
+  - fab
+---
+
+To operate machines with AC, you need to export a [`.acim`](../developer-guide/acim.md) file from your 3D modeling software. This file contains all the information required for the fabrication process, such as geometric information (e.g., joints and holes) and fabrication metadata (e.g., ID, fabrication status).
+
+![acim export image ><](../../assets/images/woodworking/draw_overview.png){width=500}
