+++ docs/augmented-woodworking/cut-chainsaw.md
@@ -1,1 +1,31 @@
-Placeholder
+---
+tags:
+  - acim
+  - UI
+  - cut
+  - chainsaw-blade
+  - fab
+---
+
+Description
+
+
+<br>
+
+# Step-by-step
+
+<!-- TODO: Select the current cut -->
+
+<!-- TODO: Intro to UI and feedback components (pos, rot, depth) -->
+
+<!-- TODO: Depth: adjust blade -->
+
+<!-- TODO: Position: placing blade -->
+
+<!-- TODO: Rotation: rotate blade -->
+
+<!-- TODO: Guide the cut towards the end -->
+
+<!-- TODO: Quick inspection to see the overlap -->
+
+<!-- TODO: Set as done the current cut -->
