### 2025-03-21 9b92d56
CAP: finished first version for circular saw section
