+++ docs/augmented-woodworking/cut-circularsaw.md
@@ -7,16 +7,17 @@ tags:
   - fab
 ---
 
-Description
+Here's the step-by-step guide to show how to use augmented-carpentry with any circular saw to realize basic joints (cross-lap, half-lap, and butt joints).
 
+
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1068168766?h=51370fb4a5&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
 
 <br>
 
+
 # Step-by-step
 
-<!-- TODO: Select the current cut -->
-
-<!-- TODO: Intro to UI and feedback components (pos, rot, depth) -->
+<!-- Intro to UI and feedback components (pos, rot, depth) -->
 <div class="sevinch-display-tutorial" markdown>
 <figure markdown>
 ![Image title](../assets/images/woodworking/cut_circular_uigen.png)
@@ -28,7 +29,7 @@ This is the overview of the feedback and widget system to assist you during the 
 
 **b** - Feedback for blade rotation (in degrees). The goal is to diminish the line length corresponding to the angle error, to zero (green).
 
-**c** - Feedback for balde depth (in mm). The goal is to to get to a value close to zero (green). In `00|35`, the first tow digit represent the distance of the projected blade to the bottom face of the cut, and the last two digits represent the closest distance from the blade to the bottom face of the cut (two purple-magenta points).
+**c** - Feedback for balde depth (in mm). The goal is to to get to a value close to zero (green). In `00|00`, the first tow digit represent the distance of the projected blade to the bottom face of the cut, and the last two digits represent the closest distance from the blade to the bottom face of the cut (two purple-magenta points).
 
 **d** - Unselected face.
 
@@ -61,7 +62,7 @@ This is the overview of the feedback and widget system to assist you during the 
 </div>
 </figure>
 <div class="description" markdown>
-Adjust first the height of the blade (**c**).
+Adjust first the height of the blade (**c**) until the indicator becomes green.
 </div>
 </div>
 
@@ -86,15 +87,197 @@ Adjust first the height of the blade (**c**).
 </div>
 </figure>
 <div class="description" markdown>
-Adjust now the position (**a**).
+Adjust now the position (**a**) until the indicator becomes green.
 </div>
 </div>
 
 
-<!-- TODO: Rotation: rotate blade -->
+<!-- Rotation: rotate blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1068047212?h=d8411fd54a&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/frame-overlay.png"
+        class="frame-overlay"
+    />
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/overlay_symbol_rotate.png"
+        class="symbol-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Adjust now the rotation (**b**) until the indicator becomes green.
+</div>
+</div>
 
-<!-- TODO: Guide the cut towards the end -->
 
[…]
