+++ docs/augmented-woodworking/cut-swordsaw.md
@@ -29,3 +29,27 @@ Description
 <!-- TODO: Quick inspection to see the overlap -->
 
 <!-- TODO: Set as done the current cut -->
+
+
+
+<!-- TODO: hint: butt joint -->
+!!! cut-joint "Butt joints"
+    <div class="sevinch-display-tutorial" markdown>
+    <figure markdown>
+    <div class="video-container">
+        <iframe 
+            src="https://player.vimeo.com/video/1066639778?h=943e283a32&amp;badge=1&amp;autopause=0&amp;player_id=0&amp;app_id=58479&amp;title=0&amp;byline=0&amp;portrait=0" 
+            frameborder="0" 
+            allow="autoplay; fullscreen; picture-in-picture" 
+            allowfullscreen>
+        </iframe>
+        <img 
+            src="https://github.com/ibois-epfl/augmented-carpentry/blob/main/docs/assets/images/frame-overlay.png?raw=true"
+            class="frame-overlay"
+        />
+    </div>
+    </figure>
+    <div class="admonition-seveinch-desc" markdown>
+    Description.
+    </div>
+    </div>
