### 2025-03-21 b7bf035
FIX-WIP: modfiication to structure section
