+++ docs/augmented-woodworking/index.md
@@ -1,3 +1,0 @@
-In this index we need to explain the general structure of the UI and the sections.
-
-HEre we should explain the how we explain and use the steps and how what is the lgend for the admonitions block (red, we cannot fix, bklue, trouble shooting, light blue is a tip etc)
