+++ docs/augmented-woodworking/cut-chainsaw.md
@@ -7,25 +7,223 @@ tags:
   - fab
 ---
 
-Description
+Here's the step-by-step guide on how to use augmented-carpentry with any circular saw to realize basic woodworking operations.
+
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1068850648?h=c7dce822c4&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
 
 
 <br>
 
 # Step-by-step
 
-<!-- TODO: Select the current cut -->
 
-<!-- TODO: Intro to UI and feedback components (pos, rot, depth) -->
+<!-- Intro to UI and feedback components (pos, rot, depth) -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/cut_chainsaw_ui.png)
+</figure>
+<div class="description" markdown>
+The assistance interface for the chainsaw is composed by the following elements:
 
-<!-- TODO: Depth: adjust blade -->
+**a** - Feedback for balde position (in cm). The goal is to diminish the line length corresponding to the position error, to zero (green).
 
-<!-- TODO: Position: placing blade -->
+**b** - Feedback for blade rotation (in degrees). The goal is to diminish the line length corresponding to the angle error, to zero (green).
 
-<!-- TODO: Rotation: rotate blade -->
+**c** - This is particular to the chainsaw feedback system. In order to obtain straight cuts both the base and tip of the chain needs to have equal heights. The yellow lines represent the distance to the bottom faces of bottom and tip blade. The idea is to have equidistant heights so that the two lines become green. The value `d:00` represents the shortest distances of both.
 
-<!-- TODO: Guide the cut towards the end -->
+**d** - The blade projection on the bottom face.
+</div>
+</div>
 
-<!-- TODO: Quick inspection to see the overlap -->
 
-<!-- TODO: Set as done the current cut -->
+<!-- Position: placing blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1068875476?h=f7f93dcf44&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/frame-overlay.png"
+        class="frame-overlay"
+    />
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/overlay_symbol_leftright.png"
+        class="symbol-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Adjust now the position (**a**) until the indicator becomes green.
+</div>
+</div>
+
+
+<!-- Rotation: rotate blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1068880690?h=95fec4aad5&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
[…]
