### 2025-03-25 db53916
CAP-CAD: chapter on miter saw
