+++ docs/augmented-woodworking/cut-swordsaw.md
@@ -7,38 +7,162 @@ tags:
   - fab
 ---
 
-Description
+Here's the step-by-step guide on how to use augmented-carpentry with a sword saw to realize basic joints: half-lap, and mainly butt joints. The feedback system is similar to the one for [chainsaws](cut-chainsaw.md).
 
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1069262039?h=ae496471bd&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
 
 <br>
 
 # Step-by-step
 
-<!-- TODO: Select the current cut -->
+<!-- Intro to UI and feedback components (pos, rot, depth) -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/cut_swordsaw_ui.png)
+</figure>
+<div class="description" markdown>
+This is the overview of the feedback and widget system to assist you during the cut.
 
-<!-- TODO: Intro to UI and feedback components (pos, rot, depth) -->
+**a** - Feedback for blade rotation (in degrees). The goal is to diminish the line length corresponding to the angle error, to zero (green).
 
-<!-- TODO: Depth: adjust blade -->
+**b** - Feedback for balde position (in cm). The goal is to diminish the line length corresponding to the position error, to zero (green).
 
-<!-- TODO: Position: placing blade -->
+**c,e** - This is particular to the chainsaw feedback system. In order to obtain straight cuts both the base and tip of the chain needs to have equal heights. The yellow lines represent the distance to the bottom faces of bottom and tip blade. The idea is to have equidistant heights so that the two lines become green. The value `d:00` represents the shortest distances of both.
 
-<!-- TODO: Rotation: rotate blade -->
+**d** - The blade projection on the bottom face.
 
-<!-- TODO: Guide the cut towards the end -->
+</div>
+</div>
 
-<!-- TODO: Quick inspection to see the overlap -->
+<!-- Position: placing blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1069266325?h=646ddee2e2&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/frame-overlay.png"
+        class="frame-overlay"
+    />
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/overlay_symbol_leftright.png"
+        class="symbol-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Adjust now the position until the indicator becomes green.
+</div>
+</div>
 
-<!-- TODO: Set as done the current cut -->
+
+<!-- Rotation: rotate blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
[…]
