### 2025-03-27 a83d205
CAP: add drilling section
