### 2025-03-28 29592e3
Documentation patch + adds (#562)
* ADD: instructions for building docs

* ADD: support for lfs gimp

* CAP-WIP: template and working on mapping section

* CAP-ADD: video template for tutorials with mask

* WIP: working on map chapter

* ADD-MILESTONE: done map section

* FIX: correct placing of iframe

* CAP: finished a version for the execution-model

* WIP: working on calibrage

* CAP: finished calib chapter

* FIX: deprecated upload artifact action

* CAP: added the rhino package for ac

* WIP: working on cut

* CAP: finished first version for circular saw section

* FIX-WIP: modfiication 
[…]
