+++ docs/augmented-woodworking/examples.md
@@ -1,0 +1,6 @@
+In this index we need to explain the general structure of the UI and the sections.
+
+HEre we should explain the how we explain and use the steps and how what is the lgend for the admonitions block (red, we cannot fix, bklue, trouble shooting, light blue is a tip etc)
+
+
+<!-- TODO: here add the 3d recordings -->
