+++ docs/augmented-woodworking/cut-mitersaw.md
@@ -1,0 +1,254 @@
+---
+tags:
+  - acim
+  - UI
+  - cut
+  - circular-blade
+  - fab
+---
+
+Here's the step-by-step guide on how to use augmented-carpentry with any miter saw to realize basic joints (cross-lap, half-lap, and butt joints). The guidance system is similar to the one for [circular saws](augmented-circular-saw.md).
+
+<div class="video-wrapper"><iframe src="https://player.vimeo.com/video/1069137536?h=eb40e565bc&amp;badge=0&amp;autopause=0&amp;title=0&amp;player_id=0&amp;app_id=58479" frameborder="0" allow="autoplay; fullscreen; picture-in-picture; clipboard-write; encrypted-media" style="position:absolute;top:0;left:0;width:100%;height:100%;"></iframe></div><script src="https://player.vimeo.com/api/player.js"></script>
+
+<br>
+
+# Step-by-step
+
+
+<!-- Intro to UI and feedback components (pos, rot, depth) -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+![Image title](../assets/images/woodworking/cut_mitersaw_ui.png)
+</figure>
+<div class="description" markdown>
+This is the overview of the feedback and widget system to assist you during the cut.
+
+**a** - Feedback for balde position (in cm). The goal is to diminish the line length corresponding to the position error, to zero (green).
+
+**b** - Feedback for blade rotation (in degrees). The goal is to diminish the line length corresponding to the angle error, to zero (green).
+
+**c** - Feedback for balde depth (in mm). The goal is to to get to a value close to zero (green). In `00|00`, the first tow digit represent the distance of the projected blade to the bottom face of the cut, and the last two digits represent the closest distance from the blade to the bottom face of the cut (two purple-magenta points).
+
+**d** - Projection of the blade thickness on the face to be cut.
+</div>
+</div>
+
+
+<!-- Rotation: rotate blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1069143962?h=2fbeababdf&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/frame-overlay.png"
+        class="frame-overlay"
+    />
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/overlay_symbol_rotate.png"
+        class="symbol-overlay"
+    />
+</div>
+</figure>
+<div class="description" markdown>
+Adjust the rotation (**b**) until the indicator becomes green.
+</div>
+</div>
+
+
+<!-- Position: placing blade -->
+<div class="sevinch-display-tutorial" markdown>
+<figure markdown>
+<div class="video-container">
+    <iframe 
+        src="https://player.vimeo.com/video/1069145307?h=e4cbec299a&amp;background=1&amp;autopause=0&amp;loop=1&amp;autoplay=1&amp;muted=1&amp;controls=0&amp;title=0&amp;byline=0&amp;portrait=0" 
+        frameborder="0" 
+        allow="autoplay; fullscreen; picture-in-picture" 
+        allowfullscreen>
+    </iframe>
+    <img 
+        src="{{ config.site_url }}/assets/images/woodworking/frame-overlay.png"
+        class="frame-overlay"
+    />
+    <img 
[…]
