### 2025-03-31 b161b19
FIX: footer for mobile and non
