+++ assets/.doxy/acdoxygen/acdoxygen/GORegistry_8h_source.md
@@ -1,0 +1,192 @@
+
+
+# File GORegistry.h
+
+[**File List**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**GOSys**](dir_5f167dfbcc50bcf5ffb2c1a7f5cf90ca.md) **>** [**GORegistry.h**](GORegistry_8h.md)
+
+[Go to the documentation of this file](GORegistry_8h.md)
+
+
+```C++
+// #####################################################################
+// >>>>>>>>>>>>>>>>>>>>> BEGINNING OF LEGAL NOTICE >>>>>>>>>>>>>>>>>>>>>
+//######################################################################
+//
+// This source file, along with its associated content, was authored by
+// Andrea Settimi, Hong-Bin Yang, Naravich Chutisilp, and numerous other
+// contributors. The code was originally developed at the Laboratory for
+// Timber Construction (IBOIS, director: Prof. Yves Weinand) at the School of 
+// Architecture, Civil and Environmental Engineering (ENAC) at the Swiss
+// Federal Institute of Technology in Lausanne (EPFL) for the Doctoral
+// Research "Augmented Carpentry" (PhD researcher: Andrea Settimi,
+// co-director: Dr. Julien Gamerro, director: Prof. Yves Weinand).
+//
+// Although the entire repository is distributed under the GPL license,
+// these particular source files may also be used under the terms of the
+// MIT license. By accessing or using this file, you agree to the following:
+//
+// 1. You may reproduce, modify, and distribute this file in accordance
+//    with the terms of the MIT license.
+// 2. You must retain this legal notice in all copies or substantial
+//    portions of this file.
+// 3. This file is provided "AS IS," without any express or implied
+//    warranties, including but not limited to the implied warranties of
+//    merchantability and fitness for a particular purpose.
+//
+// If you cannot or will not comply with the above conditions, you are
+// not permitted to use this file. By proceeding, you acknowledge and
+// accept all terms and conditions herein.
+//
+//######################################################################
+// <<<<<<<<<<<<<<<<<<<<<<< END OF LEGAL NOTICE <<<<<<<<<<<<<<<<<<<<<<<<
+// #####################################################################
+//
+#pragma once
+
+#include "AIAC/Log.h"
+#include "AIAC/GOSys/GOPrimitive.h"
+
+
+namespace AIAC
+{
+    class GORegistry
+    {
+    public:
+        GORegistry() = default;
+        ~GORegistry() = default;
+
+        void Init()
+        {
+            Clear();
+        }
+    
+        template<typename T>
+        void Register(std::shared_ptr<T> go)
+        {
+            static_assert(std::is_base_of<AIAC::GOPrimitive, T>::value, "Pushed type is not subclass of GOPrimitive!");
+            m_GOMap.emplace(go->GetId(), go);
+        }
+
+        template<typename T>
+        void Register(const uint32_t& id, std::shared_ptr<T> go)
+        {
+            static_assert(std::is_base_of<AIAC::GOPrimitive, T>::value, "Pushed type is not subclass of GOPrimitive!");
+            m_GOMap.emplace(id, go);
+        }
+
+        inline void Unregister(const uint32_t& id)
+        {
+            m_GOMap.erase(id);
+        }
+
+        template<typename T>
+        std::shared_ptr<T> GetGO(const uint32_t& id)
+        {
+            static_assert(std::is_base_of<AIAC::GOPrimitive, T>::value, "Type to get is not subclass of GOPrimitive!");
+            auto it = m_GOMap[id];
+            // check if the static pointer cast succeed
+            if (it != nullptr)
+            {
+                try
+                {
+                    return std::dynamic_pointer_cast<T>(it);
+                }
+                catch (const std::bad_cast& e)
[…]
