+++ assets/.doxy/acdoxygen/acdoxygen/GLObject_8cpp.md
@@ -1,0 +1,92 @@
+
+
+# File GLObject.cpp
+
+
+
+[**FileList**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**Render**](dir_4231f99b70cbd7a69a19f070b3954fcb.md) **>** [**GLObject.cpp**](GLObject_8cpp.md)
+
+[Go to the source code of this file](GLObject_8cpp_source.md)
+
+
+
+* `#include "GLObject.h"`
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Namespaces
+
+| Type | Name |
+| ---: | :--- |
+| namespace | [**AIAC**](namespaceAIAC.md) <br> |
+
+
+## Classes
+
+| Type | Name |
+| ---: | :--- |
+| struct | [**CylinderPole**](structAIAC_1_1CylinderPole.md) <br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+------------------------------
+The documentation for this class was generated from the following file `src/AIAC/Render/GLObject.cpp`
+

[…]
