### 2025-04-09 65ef601
ADD: photo for motivation doc page
