+++ assets/.doxy/acdoxygen/acdoxygen/Log_8h.md
@@ -1,0 +1,156 @@
+
+
+# File Log.h
+
+
+
+[**FileList**](files.md) **>** [**AIAC**](dir_21da83368f7816722f2b707a7b03c84f.md) **>** [**Log.h**](Log_8h.md)
+
+[Go to the source code of this file](Log_8h_source.md)
+
+
+
+* `#include <spdlog/spdlog.h>`
+* `#include <spdlog/sinks/stdout_color_sinks.h>`
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Namespaces
+
+| Type | Name |
+| ---: | :--- |
+| namespace | [**AIAC**](namespaceAIAC.md) <br> |
+
+
+## Classes
+
+| Type | Name |
+| ---: | :--- |
+| class | [**Log**](classAIAC_1_1Log.md) <br> |
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+## Macros
+
+| Type | Name |
+| ---: | :--- |
+| define  | [**AIAC\_CRITICAL**](Log_8h.md#define-aiac_critical)  <br> |
+| define  | [**AIAC\_ERROR**](Log_8h.md#define-aiac_error)  <br> |
+| define  | [**AIAC\_INFO**](Log_8h.md#define-aiac_info)  <br> |
+| define  | [**AIAC\_WARN**](Log_8h.md#define-aiac_warn)  <br> |
+
+## Macro Definition Documentation
+
+
+
+
+
+### define AIAC\_CRITICAL 
+
+```C++
+#define AIAC_CRITICAL 
+```
+
+
+
+
+<hr>
+
+
+
+### define AIAC\_ERROR 
+
+```C++
+#define AIAC_ERROR 
+```
+
+
+
+
+<hr>
+
+
+
+### define AIAC\_INFO 
+
+```C++
+#define AIAC_INFO 
+```
+
+
+
+
+<hr>
+
+
+
+### define AIAC\_WARN 
+
+```C++
+#define AIAC_WARN 
+```
+
+
+
+
+<hr>
+
+------------------------------
+The documentation for this class was generated from the following file `src/AIAC/Log.h`
+
