### 2025-09-13 #7686 @arfon
@9and3 – I'm sorry this one has taken so very long to get to. Could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?
