### 2025-10-24 #9296 @preetishkakkar
- [X] **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines])
