### 2026-02-08 #9296 @ApocalyVec
Hi @arfon and @9and3 — apologies for the delayed start on my end; I ran into an unexpected deadline crunch over the past couple of weeks. I’m starting my review now and will complete my checklist and feedback by Feb 13. Thanks for your patience.
