### 2026-02-17 #9296 @ApocalyVec
Hi @arfon and @9and3 — quick update: I started going through the README, but the online documentation link appears to be broken (it points to missinglink: https://github.com/ibois-epfl/augmented-carpentry/blob/main/missinglink
). Could you please update this to the correct docs URL? I’ll continue and finalize the review once the documentation link is fixed.
