### 2026-02-17 #9296 @9and3
Hello @arfon , thanks for taking the time 👐
The broken will be fixed by the end of the day (opened a PR).
In any case the docs is here: [ibois-epfl.github.io/augmented-carpentry/](https://ibois-epfl.github.io/augmented-carpentry/)
