### 2026-03-30 ibois-epfl/augmented-carpentry#569 @preetishkakkar
Docs: DEPLOY_ON_TOUCH is ON by default causing error when it's build for laptop
Part of review for https://github.com/openjournals/joss-reviews/issues/9296

**Description**

The documentation (https://ibois-epfl.github.io/augmented-carpentry/installation/#config-and-build-wearable) doesn't say that for laptop you have to set  DEPLOY_ON_TOUCH to off, this causes the following error when the application is run.

[source TouchMonitor.cpp] [function Init] [line 123] [19:47:40] APP: No resolution specified found in list monitors.

**Proposed change (optional)**
Update the laptop build to be explicit about setting DEPLOY_ON_TOUCH to OFF
