### 2026-03-30 #9296 @preetishkakkar
Hello @9and3 , I don’t currently have the setup needed to properly test the app on my machine, so I’m unable to verify its behavior end-to-end. I’m also not sure whether the instructions here are sufficient: https://ibois-epfl.github.io/augmented-carpentry/augmented-woodworking/map/

Is there a way to simulate the camera input, perhaps using a virtual camera or similar approach so I can test the app more easily in a local environment?
