### 2026-04-04 #9296 @9and3
Hello @preetishkakkar , thanks for the review!

Augmented Carpentry is an open software made for open fabrication. The instructions you are mentioning are complemented by an exhaustive description of hardware setup, here: https://ibois-epfl.github.io/augmented-carpentry/hardware/ .

The software could be tested with a very generic webcam readable from a USB port. We dedicated a chapter to the camera sensor, including the calibration process: https://ibois-epfl.github.io/augmented-carpentry/hardware/sensors/.
