### 2026-04-04 ibois-epfl/augmented-carpentry#569 @9and3
Thank you for flagging this documentation gap @preetishkakkar !
The primary target device for Augmented Carpentry is a touch display, so `DEPLOY_ON_TOUCH=ON` is the intended default for our main use case. Setting it to OFF by default would actually be counterproductive for most users deploying on the target hardware.

That said, you're absolutely right that the laptop build documentation should explicitly mention this flag. We mention this already in our docs README.md:
