### 2026-04-28 #9296 @preetishkakkar
Hi @arfon — I’m working on this right now, but it’s a bit tricky to test because the setup requires printing stickers and recreating what’s shown in the documentation (https://ibois-epfl.github.io/augmented-carpentry/augmented-woodworking/map/).

It would really help if there were a way to simulate the workflow with a virtual camera similar to what’s demonstrated in the docs.

@9and3 — once I’ve calibrated the webcam, would you recommend testing by printing the tags and placing them on a table or some other surface? How do you suggest we simulate these workflows without access to an actual timber piece?
