### 2026-04-29 #9296 @9and3
Hello @preetishkakkar , you can just print them and test them on a surface it is fine just for a smoke run. Augmented Carpentry is a free and open source software dedicated to timber construction so it is understandable that you might not replicate every step without actually going into a fabrication process as indicated in the documentation https://ibois-epfl.github.io/augmented-carpentry/ . Thanks for the reviewing effort!
